+++ src/game.ts
@@ -1,13 +1,12 @@
-// game.ts
 // Copyright (c) 2024 KibaOfficial
-//
+// 
 // This software is released under the MIT License.
 // https://opensource.org/licenses/MIT
 
 import { DeltaTime } from "./classes/DeltaTime.js";
 import { InputManager } from "./classes/InputManager.js";
 import { WindowManager } from "./classes/WindowManager.js";
-import { getGameInit, getGameRun, setGameTitle } from "./constants.js";
+import { getGameInit, getGameRun } from "./constants.js";
 import { initGame, updateFPS } from "./utils.js";
 import { drawPauseOverlay } from "./Graphics.js";
 import { SceneManager } from "./classes/SceneManager.js";
@@ -17,21 +16,32 @@ import { Scene } from "./classes/Scene.js";
 const deltaTime = new DeltaTime();
 const inputManager = new InputManager();
 const winManager = new WindowManager();
-const sceneManager = new SceneManager();
+const sceneManager = new SceneManager(winManager);
 
-// Create and add the gameScene to the SceneManager
-const gameScene = new Scene("gameScene", window.innerWidth, window.innerHeight);
+// Create and add the gameScene to the SceneManager with a specific background color
+const gameScene = new Scene(
+  "gameScene",
+  window.innerWidth,
+  window.innerHeight,
+  "#87CEEB", // Sky Blue
+  1 // Lower z-index
+);
 sceneManager.addScene(gameScene);
-sceneManager.switchScene("gameScene");
+sceneManager.activateScene("gameScene");
 
-// Set gametitle
-setGameTitle("Kio T Engine");
+// Create and add the uiScene to the SceneManager
+const uiScene = new Scene(
+  "uiScene",
+  window.innerWidth,
+  window.innerHeight,
+  "transparent", // Transparent background
+  2 // Higher z-index
+);
+sceneManager.addScene(uiScene);
+sceneManager.activateScene("uiScene");
 
 // Initialize the game loop
 export function gameLoop(currentTime: DOMHighResTimeStamp) {
-  // Resize the canvas using WindowManager
-  winManager.resizeCanvas(gameScene.canvas);
-
   // Check if the game is running
   if (getGameRun()) {
     if (!getGameInit()) {
@@ -49,12 +59,11 @@ export function gameLoop(currentTime: DOMHighResTimeStamp) {
     const dt = deltaTime.getDelta();
 
     // Update the FPS display
-    updateFPS(dt, gameScene.ctx);
+    updateFPS(dt, uiScene.ctx);
 
     // Call the update and render methods of the current scene
-    // but not needed yet
-    // sceneManager.update(dt);
-    // sceneManager.render();
+    sceneManager.update(dt);
+    sceneManager.render();
 
     requestAnimationFrame(gameLoop);
   }