+++ src/utils.ts
@@ -3,7 +3,7 @@
 // This software is released under the MIT License.
 // https://opensource.org/licenses/MIT
 
-import { getGameTitle, setGameInit } from "./constants.js";
+import { getGameTitle, setGameInit, setGameTitle } from "./constants.js";
 
 export function updateLayer(
   ctx: CanvasRenderingContext2D | null,
@@ -17,18 +17,20 @@ export function updateLayer(
 }
 
 export function initGame(): void {
+  // Set gametitle
+  setGameTitle("Kio T Engine");
   document.title = getGameTitle();
   setGameInit(true);
 }
 
 export function updateFPS(dt: number, ctx: CanvasRenderingContext2D) {
-  const fps = 1 / dt
-  
+  const fps = 1 / dt;
+
   updateLayer(ctx, (ctx: CanvasRenderingContext2D) => {
     ctx.font = '20px Arial';
     ctx.fillStyle = 'black';
     ctx.textAlign = 'right';
     ctx.textBaseline = 'top';
     ctx.fillText(`FPS: ${Math.round(fps)}`, ctx.canvas.width - 10, 10);
-  })
+  });
 }